Changes... fix some multiplayer stuff and convert a widget to come from C++ and listen to events this way

+++ Source/StickyBombTest/StickyBombPlayerState.h
@@ -6,7 +6,7 @@
 #include "GameFramework/PlayerState.h"
 #include "StickyBombPlayerState.generated.h"
 
-DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnAmmoCountChanged, float, NewAmmo);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_OneParam(FOnAmmoCountChanged, int, NewAmmo);
 
 /**
  *
